steal a nice color palette
https://github.com/Gnuplotting/gnuplot-palettes/blob/master/turbo.pal

+++ color.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-uint32_t rgb(float);
+uint32_t rgb(unsigned);
 uint32_t gray(float);
 
 #endif
